move log level to logging
